Refactor and moved sidebar components, files inside a sidebar folder to organize them by sidebar category
