+++ app/(documents)/_components/sidebar/constants.ts
@@ -1,0 +1,4 @@
+const MAX_WIDTH = 450;
+const MIN_WIDTH = 250;
+
+export { MAX_WIDTH, MIN_WIDTH };
