+++ app/(documents)/_components/sidebar/side-bar-menu.tsx
@@ -2,13 +2,15 @@
 
 import { AlignJustify } from "lucide-react";
 
-import { useSideBar } from "../zustand-store/sidebar-store";
 import { cn } from "@/lib/utils";
+
+import { useSideBar } from "../../zustand-store/sidebar-store";
 import { useMediaQuery } from "usehooks-ts";
 
 export const SideBarMenu = () => {
   const { setIsCollapsed, setIsTrasitioning, isCollapsed } = useSideBar();
   const isMobile = useMediaQuery("(max-width: 768px)");
+
   return (
     <AlignJustify
       onClick={() => {
